requirements.txt
flet>=0.24.0,<0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
telebot>=4.0.0
shelve